tests/wpt/web-platform-tests/2dcontext/hit-regions/hitregions-members-exist.html

<!DOCTYPE html>
<title>Canvas test: 2d.hitregions.members.exist</title>
<meta name="author" content="Constantine Kim">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>

<h1>2d.hitregions.members.exist</h1>
<canvas></canvas>
<div id="log"></div>
<script>
test(function () {
  var ctx = document.createElement('canvas').getContext('2d');
  assert_equals(typeof ctx.addHitRegion, 'function');
}, 'context.addHitRegion Exists');

test(function () {
  var ctx = document.createElement('canvas').getContext('2d');
  assert_equals(typeof ctx.removeHitRegion, 'function');
}, 'context.removeHitRegion Exists');

test(function () {
  var ctx = document.createElement('canvas').getContext('2d');
  assert_equals(typeof ctx.clearHitRegions, 'function');
}, 'context.clearHitRegions Exists');

</script>
</body>
</html>
